requirements.txt
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
distlib==0.3.2
Django==3.2.4
djangorestframework==3.12.4
filelock==3.0.12
mysqlclient==2.0.3
pipenv==2021.5.29
protobuf==3.17.3
pycodestyle==2.7.0
pytz==2021.1
pywatchman==1.4.1
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.7
virtualenv-clone==0.5.4
